Refactor Validation Tests a Bit
