+++ tests/tests/binding_array/mod.rs
@@ -2,4 +2,3 @@ mod buffers;
 mod sampled_textures;
 mod samplers;
 mod storage_textures;
-mod validation;
