+++ tests/validation_tests/api/mod.rs
@@ -1,2 +1,4 @@
+mod binding_arrays;
 mod buffer;
+mod buffer_slice;
 mod texture;
