+++ tests/validation_tests/root.rs
@@ -6,6 +6,10 @@ mod noop;
 /// Obtain a device using [`wgpu::Backend::Noop`].
 /// This should never fail.
 fn request_noop_device() -> (wgpu::Device, wgpu::Queue) {
+    request_noop_device_with_desc(&wgpu::DeviceDescriptor::default())
+}
+
+fn request_noop_device_with_desc(desc: &wgpu::DeviceDescriptor) -> (wgpu::Device, wgpu::Queue) {
     let instance = wgpu::Instance::new(&wgpu::InstanceDescriptor {
         backends: wgpu::Backends::NOOP,
         backend_options: wgpu::BackendOptions {
@@ -20,6 +24,5 @@ fn request_noop_device() -> (wgpu::Device, wgpu::Queue) {
             .expect("adapter");
     assert_eq!(adapter.get_info().backend, wgpu::Backend::Noop);
 
-    pollster::block_on(adapter.request_device(&wgpu::DeviceDescriptor::default(), None))
-        .expect("device")
+    pollster::block_on(adapter.request_device(desc, None)).expect("device")
 }
